Remove packer template
